Identify panel break out
