+++ packages/base/src/features/identify/types/editorTypes.ts
@@ -1,0 +1,17 @@
+export interface IPropertyEditorState {
+  editingFeatureIndex: number | null;
+  editorMode: 'add' | 'edit' | null;
+  editingPropertyKey: string | null;
+  newPropertyKey: string;
+  newPropertyValue: string;
+  isSavingProperty: boolean;
+}
+
+export interface IPropertyEditorActions {
+  onEditProperty: (rowIndex: number, propertyKey: string, value: any) => void;
+  onStartAddProperty: (rowIndex: number) => void;
+  onSaveProperty: (feature: any, rowIndex: number) => void;
+  onCancelProperty: () => void;
+  onNewPropertyKeyChange: (value: string) => void;
+  onNewPropertyValueChange: (value: string) => void;
+}
